LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/runtime-error-same-origin-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: undefined_variable

PASS window.onerror - runtime error in <script src=...>
PASS window.onerror - runtime error in <script src=...> (column)
